requirements.txt
# Core dependencies
click>=8.1.0
python-dotenv>=1.0.0
rich>=13.0.0
pydantic>=2.0.0
loguru>=0.7.0
psutil>=5.9.0

# AWS support
boto3>=1.34.0
botocore>=1.34.0

# Data processing
pandas>=2.0.0
numpy<2,>=1.24.0

# Video processing
moviepy==1.0.3
ffmpeg-python>=0.2.0
sentence-transformers>=2.2.2
lancedb>=0.6.12

# HTTP requests
requests>=2.31.0

# JSON processing
orjson>=3.9.0

# AI/ML dependencies
openai>=1.0.0
google-generativeai>=0.3.0
google-genai>=1.0.0  # New unified Gemini SDK
google-cloud-aiplatform>=1.35.0
openai-whisper>=20231117
faster-whisper>=1.0.0

# YouTube API dependencies
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0

# Visualization (for analysis script)
matplotlib>=3.7.0
seaborn>=0.12.0

# Computer vision (for clip layout analysis)
opencv-python<4.12,>=4.8.0
ultralytics>=8.2.0,<9

# Phase 3: Container Modularity
# (click and python-dotenv already included above) 